POR-2613: move employee search/filtering to a util file
