+++ src/components/employees/tags/TagsTable.vue
@@ -54,7 +54,7 @@
           v-model="editedTag.employees"
           :disabled="tagLoading"
           :items="filteredEmployees"
-          :custom-filter="customFilter"
+          :custom-filter="employeeFilter"
           multiple
           variant="underlined"
           chips
@@ -128,6 +128,7 @@ import { generateUUID, isMobile } from '@/utils/utils';
 import { firstAndLastName, fullName, nicknameAndLastName } from '@/shared/employeeUtils';
 import { getRequiredRules } from '@/shared/validationUtils';
 import { updateStoreExpenseTypes, updateStoreTags } from '@/utils/storeUtils';
+import { employeeFilter } from '@/shared/filterUtils';
 
 import DeleteModal from '@/components/modals/DeleteModal.vue';
 import { AxiosError } from 'axios';
@@ -193,29 +194,6 @@ async function createTag() {
   this.editedTag = null;
   this.creatingTag = false;
 } // createTag
-
-/**
- * Custom filter for employee autocomplete options.
- *
- * @param _ - unused
- * @param queryText - text used for filtering
- * @param item - employee
- * @return string - filtered employee name
- */
-function customFilter(_, queryText, item) {
-  item = item.raw;
-
-  const query = queryText ? queryText : '';
-  const nickNameFullName = item.nickname ? `${item.nickname} ${item.lastName}` : '';
-  const firstNameFullName = `${item.firstName} ${item.lastName}`;
-
-  const queryNickName = nickNameFullName.toString().toLowerCase().indexOf(query.toString().toLowerCase());
-  const queryFirstName = firstNameFullName.toString().toLowerCase().indexOf(query.toString().toLowerCase());
-
-  if (queryNickName >= 0) return queryNickName;
-  if (queryFirstName >= 0) return item.nickname ? true : queryFirstName;
-  return false;
-} // customFilter
 
 /**
  * Deletes a tag.
@@ -349,12 +327,7 @@ function tableFilter(__, search, item) {
   if (item.tagName.toLowerCase().includes(lcSearch)) return true; // early exit if tag name matches search
   _.forEach(item.employees, (id) => {
     let e = _.find(this.$store.getters.employees, (emp) => emp.id === id);
-    if (
-      e &&
-      (this.nicknameAndLastName(e).toLowerCase().includes(lcSearch) ||
-        this.firstAndLastName(e).toLowerCase().includes(lcSearch) ||
-        this.fullName(e).toLowerCase().includes(lcSearch))
-    ) {
+    if (this.employeeFilter(null, search, e)) {
       found = true;
     }
   });
@@ -464,11 +437,11 @@ export default {
   methods: {
     cancelEdit,
     createTag,
-    customFilter,
     deleteTag,
     displayError,
     displaySuccess,
     editTag,
+    employeeFilter,
     emit,
     firstAndLastName,
     fullName,
