+++ src/components/reports/ReportsCertifications.vue
@@ -6,7 +6,7 @@
           <v-autocomplete
             id="employeesSearch"
             v-model="search"
-            :customFilter="customEmployeeFilter"
+            :custom-filter="employeeFilter"
             :items="employees"
             label="Search By Employee Name"
             variant="underlined"
@@ -108,8 +108,9 @@
 <script>
 import _ from 'lodash';
 import { userRoleIsAdmin, userRoleIsManager } from '@/utils/utils';
-import { customEmployeeFilter, getActive, getFullName, populateEmployeesDropdown } from './reports-utils';
+import { getActive, getFullName, populateEmployeesDropdown } from './reports-utils';
 import { getTodaysDate, isSameOrBefore } from '@/shared/dateUtils';
+import { employeeFilter } from '@/shared/filterUtils';
 
 // |--------------------------------------------------|
 // |                                                  |
@@ -386,8 +387,8 @@ export default {
   },
   methods: {
     buildCertificationsColumns,
-    customEmployeeFilter,
     chipColor,
+    employeeFilter,
     getActive,
     getFullName,
     handleClick,
