+++ src/components/reports/ReportsSecurityInfo.vue
@@ -6,7 +6,7 @@
           <v-autocomplete
             id="employeesSearch"
             v-model="search"
-            :customFilter="customEmployeeFilter"
+            :custom-filter="employeeFilter"
             :items="employees"
             label="Search By Employee Name"
             variant="underlined"
@@ -126,8 +126,9 @@
 <script>
 import _ from 'lodash';
 import { userRoleIsAdmin, userRoleIsManager } from '@/utils/utils';
+import { employeeFilter } from '@/shared/filterUtils';
 import { add, format, getTodaysDate } from '@/shared/dateUtils';
-import { customEmployeeFilter, getActive, getFullName, populateEmployeesDropdown } from './reports-utils';
+import { getActive, getFullName, populateEmployeesDropdown } from './reports-utils';
 
 // |--------------------------------------------------|
 // |                                                  |
@@ -507,8 +508,8 @@ export default {
     };
   },
   methods: {
-    customEmployeeFilter,
     chipColor,
+    employeeFilter,
     getActive,
     getBadgeExpiration,
     getClearanceType,
