+++ src/components/reports/ReportsTechnologies.vue
@@ -6,7 +6,7 @@
           <v-autocomplete
             id="employeesSearch"
             v-model="search"
-            :customFilter="customEmployeeFilter"
+            :custom-filter="employeeFilter"
             :items="employees"
             label="Search By Employee Name"
             variant="underlined"
@@ -114,7 +114,8 @@
 <script>
 import _ from 'lodash';
 import { userRoleIsAdmin, userRoleIsManager } from '@/utils/utils';
-import { customEmployeeFilter, getActive, getFullName, populateEmployeesDropdown } from './reports-utils';
+import { employeeFilter } from '@/shared/filterUtils';
+import { getActive, getFullName, populateEmployeesDropdown } from './reports-utils';
 
 // |--------------------------------------------------|
 // |                                                  |
@@ -414,8 +415,8 @@ export default {
   },
   methods: {
     buildTechnologiesColumns,
-    customEmployeeFilter,
     chipColor,
+    employeeFilter,
     getActive,
     getFullName,
     handleClick,
